math: fix float->int functions

+++ src/math/ilogbf.c
@@ -10,7 +10,7 @@ static struct f_i t[] = {
 int main(void)
 {
 	#pragma STDC FENV_ACCESS ON
-	int yi;
+	long long yi;
 	float d;
 	int e, i, err = 0;
 	struct f_i *p;
@@ -32,7 +32,7 @@ int main(void)
 			err++;
 		}
 		if (yi != p->i) {
-			printf("%s:%d: %s ilogbf(%a) want %lld got %d\n",
+			printf("%s:%d: %s ilogbf(%a) want %lld got %lld\n",
 				p->file, p->line, rstr(p->r), p->x, p->i, yi);
 			err++;
 		}
