+++ src/math/lrint.c
@@ -10,7 +10,7 @@ static struct d_i t[] = {
 int main(void)
 {
 	#pragma STDC FENV_ACCESS ON
-	int yi;
+	long long yi;
 	float d;
 	int e, i, err = 0;
 	struct d_i *p;
@@ -32,7 +32,7 @@ int main(void)
 			err++;
 		}
 		if (yi != p->i) {
-			printf("%s:%d: %s lrint(%a) want %lld got %d\n",
+			printf("%s:%d: %s lrint(%a) want %lld got %lld\n",
 				p->file, p->line, rstr(p->r), p->x, p->i, yi);
 			err++;
 		}
